backend/requirements.txt
fastapi
uvicorn
python-dotenv
sqlalchemy
pymysql
boto3
openai
pdfplumber
python-multipart
langchain
langchain-openai
faiss-cpu
pypdf
